refactor: extract EHentaiResponse and move models to model package
