+++ komf-core/src/commonMain/kotlin/snd/komf/providers/ehentai/model/EHentaiBook.kt
@@ -1,4 +1,4 @@
-package snd.komf.providers.ehentai
+package snd.komf.providers.ehentai.model
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
