+++ komf-core/src/commonMain/kotlin/snd/komf/providers/ehentai/model/EHentaiResponse.kt
@@ -1,0 +1,10 @@
+package snd.komf.providers.ehentai.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class EHentaiResponse(
+    val gmetadata: List<EHentaiBook> = emptyList(),
+    val gid: Int? = null,
+    val error: String? = null
+)
